deleted MatchService added getFirstMessage method
