--- src/main/java/ch/uzh/ifi/hase/soprafs22/rest/dto/chat/ChatCreationPostDTO.java
@@ -1,14 +1,0 @@
-package ch.uzh.ifi.hase.soprafs22.rest.dto.chat;
-
-public class ChatCreationPostDTO {
-
-  private long userId;
-
-  public long getUserId() {
-    return userId;
-  }
-
-  public void setUserId(long userId) {
-    this.userId = userId;
-  }
-}
