+++ src/main/java/ch/uzh/ifi/hase/soprafs22/service/UserService.java
@@ -502,6 +502,45 @@ public class UserService {
     return false;
   }
 
+    /**
+     * get matches of a user
+     * @param user: user from which the matches are taken
+     * @return Lsit of Matches
+     */
+    public List<Match> getMatches(User user){
+        Set<Long> matchesOfUser = user.getMatches();
+        List<Match> matches = new ArrayList<>();
+        for(Long matchId : matchesOfUser){
+            matches.add(matchRepository.getOne(matchId));
+        }
+
+        return matches;
+    }
+
+    /**
+     * Get all users which match with the known user
+     * @param user: known user
+     * @param matches: all matches from suer
+     * @return list of users which mach with known user
+     */
+    public List<User> getUsersFromMatches(User user, List<Match> matches) {
+        List<User> matchedUsers = new ArrayList<>();
+
+        // map the users to the matches
+        for (Match match : matches){
+            Pair<User,User> users = match.getUserPair();
+            // add other user (by comparing it with the user, which is known)
+            if(user.equals(users.getObj1())){
+                matchedUsers.add(users.getObj2());
+            }else{
+                matchedUsers.add(users.getObj1());
+            }
+        }
+
+        return matchedUsers;
+    }
+
+
   /**
    * Deletes a User from Repo By the USer id
    * @param userId: userId of a user
